Moved test, checked for read preference in CommandStartedEvent
